Moved body of RegistroClases to a new component.
