+++ FrontEnd/src/Pages/RegistroClasesAlumno/RegistroClasesAlumno.jsx
@@ -3,17 +3,10 @@ import React, { useState, useEffect } from 'react';
 import CircularProgress from '@mui/material/CircularProgress';
 import {
   Alert, Button, Link, AlertTitle,
-  Card,
-  CardContent,
   Typography,
-  TextField,
-  MenuItem,
 } from '@mui/material';
 import Snackbar from '@mui/material/Snackbar';
-import Autocomplete from '@mui/material/Autocomplete';
-import { DataGrid } from '@mui/x-data-grid';
 import Modal from '@mui/material/Modal';
-import SearchIcon from '@mui/icons-material/Search';
 import { useAuth0 } from '@auth0/auth0-react';
 import moment from 'moment-timezone';
 import { getStudents } from '../../api/students';
@@ -30,17 +23,14 @@ import {
 } from '../../api/waitList';
 import { findTerm, getPeriodos } from '../../api/term';
 import ConfirmationDialog from '../../Components/Dialog/ConfirmationDialog';
-import ClaseModal from '../../Components/Clase/ClaseModal';
-import MiRegistro from '../../Components/Registro/MiRegistro';
-import Clase from '../../Components/Clase/Clase';
 import { startDateDict, endDateDict } from '../../utils/constants';
 import {
-  getNivel, getHorario, getProfesor, getCupo, compararFecha,
+  compararFecha,
 } from '../../utils/utilFunctions';
 import RegistroClasesHeader from '../../Components/Registro/RegistroClasesHeader';
+import RegistroClasesBody from '../../Components/Registro/RegistroClasesBody';
 
 function RegistroClasesAlumnos({ changeContent }) {
-  const [items, setItems] = useState([]);
   const [students, setStudents] = useState(null);
   const [currentStudent, setCurrentStudent] = useState(null);
   const [error, setError] = useState(false);
@@ -48,8 +38,6 @@ function RegistroClasesAlumnos({ changeContent }) {
   const [classNames, setClassNames] = useState([]);
   const [claseRegistrada, setClaseRegistrada] = useState([]);
   const [openConfirmationDialog, setOpenConfirmationDialog] = useState(false);
-  const [openMoreInfo, setOpenMoreInfo] = useState(false);
-  const [currentClase, setCurrentClase] = useState();
   const [selectAlertOpen, setSelectAlertOpen] = useState(false);
   const [filteredClasses, setFilteredClasses] = useState(null);
   const [dialogAction, setDialogAction] = useState('');
@@ -98,95 +86,6 @@ function RegistroClasesAlumnos({ changeContent }) {
     getStudentClasses();
   }, []);
 
-  const columns = [
-    {
-      field: 'clavePeriodo',
-      headerName: 'Periodo',
-      width: 110,
-      editable: false,
-    },
-    {
-      field: 'clave',
-      headerName: 'Clave',
-      width: 90,
-    },
-    {
-      field: 'nombre_curso',
-      headerName: 'Curso',
-      width: 90,
-      editable: false,
-    },
-    {
-      field: 'nivel',
-      headerName: 'Nivel',
-      width: 100,
-      editable: false,
-      valueGetter: getNivel,
-    },
-    {
-      field: 'area',
-      headerName: 'Area',
-      width: 110,
-      editable: false,
-    },
-    {
-      field: 'modalidad',
-      headerName: 'Modalidad',
-      width: 110,
-      editable: false,
-    },
-    {
-      field: 'horario',
-      headerName: 'Horario',
-      width: 150,
-      editable: false,
-      valueGetter: getHorario,
-    },
-    {
-      field: 'profesor',
-      headerName: 'Profesor',
-      width: 140,
-      editable: 'false',
-      valueGetter: getProfesor,
-    },
-    {
-      field: 'cupos',
-      headerName: '% curso lleno',
-      width: 100,
-      editable: 'false',
-      valueGetter: getCupo,
-    },
-    {
-      field: 'actions',
-      headerName: 'Inscripción',
-      type: 'actions',
-      width: 115,
-      renderCell: (params) => (Number(params.row.cupo_actual) < Number(params.row.cupo_maximo) ? (
-        <Button
-          size="small"
-          onClick={() => handleClick(params.row)}
-          variant="outlined"
-        >
-          {params.row.status === 'Inscrito'
-            && params.row.status !== 'ListaEspera'
-            ? 'Cancelar Registro'
-            : 'Inscribir'}
-        </Button>
-      ) : (
-        <Button
-          size="small"
-          onClick={() => handleClick(params.row)}
-          variant="outlined"
-        >
-          {params.row.status === 'ListaEspera'
-            && params.row.status !== 'Inscrito'
-            ? 'Salir de Lista'
-            : 'Lista Espera'}
-        </Button>
-      )),
-    },
-  ];
-
   const handleClick = (clase) => {
     if (currentStudent == null) {
       setSelectAlertOpen(true);
@@ -208,11 +107,6 @@ function RegistroClasesAlumnos({ changeContent }) {
         );
         handleOpenDialog(clase);
     }
-  };
-
-  const handleMoreInfo = (clase) => {
-    setCurrentClase(clase);
-    setOpenMoreInfo(!openMoreInfo);
   };
 
   const handleListaEspera = async (clase) => {
@@ -345,7 +239,7 @@ function RegistroClasesAlumnos({ changeContent }) {
 
   return (
     <Box>
-      <RegistroClasesHeader 
+      <RegistroClasesHeader
         classNames={classNames}
         clases={clases}
         currentStudent={currentStudent}
@@ -353,191 +247,11 @@ function RegistroClasesAlumnos({ changeContent }) {
         setFilteredClasses={setFilteredClasses}
         setCurrentStudent={setCurrentStudent}
       />
-      <Box
-        sx={{
-          textAlign: 'center',
-          width: '100%',
-          paddingX: '20px',
-          paddingBottom: '10px',
-          overflowY: 'scroll',
-          display: { xs: 'block', md: 'none' },
-        }}
-      >
-        {filteredClasses.length !== 0 ? (
-          filteredClasses.map((e) => (
-            <Clase
-              handleClick={handleClick}
-              handleMoreInfo={handleMoreInfo}
-              key={e._id}
-              clase={e}
-            />
-          ))
-        ) : (
-          <Box
-            sx={{
-              height: '100vh',
-              display: 'flex',
-              alignContent: 'center',
-              justifyContent: 'center',
-              flexWrap: 'wrap',
-            }}
-          >
-            <Typography variant="h3" component="div" textAlign="center">
-              No hay clases disponibles por el momento.
-            </Typography>
-          </Box>
-        )}
-      </Box>
-      <Box
-        sx={{
-          display: { xs: 'none', md: 'flex' },
-          justifyContent: 'center',
-          alignItems: 'center',
-          flexDirection: 'column',
-        }}
-      >
-        <Box sx={{ display: 'flex' }}>
-          <Card
-            sx={{
-              textAlign: 'center',
-              ml: 1,
-              my: 2,
-              display: 'flex',
-            }}
-          >
-            <SearchIcon
-              color="primary"
-              width="2em"
-              height="2em"
-              sx={{ alignSelf: 'center', ml: 0.5 }}
-            />
-            <CardContent
-              sx={{
-                display: 'flex',
-                flexDirection: 'row',
-                flexWrap: 'wrap',
-                alignItems: 'center',
-                '& .MuiTextField-root': { m: 1, width: '25ch' },
-                p: 1,
-              }}
-            >
-              <Autocomplete
-                disablePortal
-                options={classNames}
-                onChange={(e, newvalue) => {
-                  setItems([
-                    {
-                      columnField: 'nombre_curso',
-                      operatorValue: 'contains',
-                      value: newvalue,
-                    },
-                  ]);
-                }}
-                onInputChange={(e, newvalue) => {
-                  setItems([
-                    {
-                      columnField: 'nombre_curso',
-                      operatorValue: 'contains',
-                      value: newvalue,
-                    },
-                  ]);
-                }}
-                renderInput={(params) => (
-                  <TextField {...params} label="Curso" />
-                )}
-              />
-              <TextField
-                style={{ fontFamily: 'arial' }}
-                label="Nivel"
-                onChange={(e) => {
-                  setItems([
-                    {
-                      columnField: 'nivel',
-                      operatorValue: 'contains',
-                      value: e.target.value,
-                    },
-                  ]);
-                }}
-                select
-              >
-                {[
-                  '',
-                  'Desde cero',
-                  'Con bases',
-                  'Intermedio',
-                  'Avanzado',
-                ].map((e) => (
-                  <MenuItem value={e} key={e}>
-                    {e}
-                  </MenuItem>
-                ))}
-              </TextField>
-              <TextField
-                style={{ fontFamily: 'arial' }}
-                label="Periodo"
-                onChange={(e) => {
-                  setItems([
-                    {
-                      columnField: 'clavePeriodo',
-                      operatorValue: 'contains',
-                      value: e.target.value,
-                    },
-                  ]);
-                }}
-              />
-              <TextField
-                style={{ fontFamily: 'arial' }}
-                label="Modalidad"
-                onChange={(e) => {
-                  setItems([
-                    {
-                      columnField: 'modalidad',
-                      operatorValue: 'contains',
-                      value: e.target.value,
-                    },
-                  ]);
-                }}
-              />
-            </CardContent>
-          </Card>
-          <MiRegistro />
-        </Box>
-        <Box
-          sx={{
-            m: 2,
-            display: 'flex',
-            width: '90%',
-            height: 600,
-            minWidth: '548px',
-            '& .theme--ListaEspera': {
-              bgcolor: 'lightyellow',
-            },
-            '& .theme--Inscrito': {
-              bgcolor: 'lightgreen',
-            },
-          }}
-        >
-          <DataGrid
-            sx={{ flexGrow: 1 }}
-            rows={filteredClasses}
-            columns={columns}
-            disableSelectionOnClick
-            getRowId={(row) => row._id}
-            getRowHeight={() => 'auto'}
-            filterModel={{
-              items,
-            }}
-            getRowClassName={(params) => `theme--${params.row.status}`}
-          />
-        </Box>
-        <Modal
-          open={openMoreInfo}
-          onClose={() => setOpenMoreInfo(!openMoreInfo)}
-          sx={{ overflowY: 'scroll' }}
-        >
-          <ClaseModal clase={currentClase} />
-        </Modal>
-      </Box>
+      <RegistroClasesBody
+        handleClick={handleClick}
+        filteredClasses={filteredClasses}
+        classNames={classNames}
+      />
       <ConfirmationDialog
         action={dialogAction}
         clase={claseRegistrada}
